Components are now more clear
Queries got separated in custom hooks
